Fix integer overflow during file length calculation on 32bit targets.

+++ src/stub.rs
@@ -75,6 +75,6 @@ impl MmapInner {
     }
 }
 
-pub fn file_len(file: &File) -> io::Result<usize> {
-    Ok(file.metadata()?.len() as usize)
+pub fn file_len(file: &File) -> io::Result<u64> {
+    Ok(file.metadata()?.len())
 }
